bit -> smallint
PostgreSQL doesn't have a `bit` data type, but it does have a `bool`.
However porting the existing SQL (especially the database inserts) to
use the `bool` type is a lot more work than is necessary. In particular
all the literals like `0` or `1` would have to be quoted. Using
`smallint` is sufficient for now.